# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_comitanigiacomo_kanso-sync-engine_internal_core_domain.Habit:
    properties:
      archived_at:
        type: string
      color:
        type: string
      created_at:
        type: string
      current_streak:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      frequency_type:
        type: string
      icon:
        type: string
      id:
        type: string
      interval:
        type: integer
      longest_streak:
        type: integer
      reminder_time:
        type: string
      sort_order:
        type: integer
      start_date:
        type: string
      target_value:
        type: integer
      title:
        type: string
      type:
        type: string
      unit:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      version:
        type: integer
      weekdays:
        items:
          type: integer
        type: array
    type: object
...
  internal_adapters_handler_http.createHabitRequest:
    properties:
      color:
        type: string
      description:
        type: string
      frequency_type:
        type: string
      icon:
        type: string
      interval:
        type: integer
      reminder_time:
        type: string
      target_value:
        type: integer
      title:
        type: string
      type:
        type: string
      unit:
        type: string
      weekdays:
        items:
          type: integer
        type: array
    required:
    - title
    type: object
...
  internal_adapters_handler_http.updateHabitRequest:
    properties:
      color:
        type: string
      description:
        type: string
      frequency_type:
        type: string
      icon:
        type: string
      interval:
        type: integer
      reminder_time:
        type: string
      target_value:
        type: integer
      title:
        type: string
      type:
        type: string
      unit:
        type: string
      version:
        type: integer
      weekdays:
        items:
          type: integer
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: Giacomo Comitani
    url: https://github.com/comitanigiacomo
  description: |-
    Backend for the Kanso Habit Tracker application.
    Features include offline-first sync, streak calculation, and JWT authentication.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kanso Sync Engine API
  version: "1.0"
paths:
...
  /habits:
    get:
      description: Get all active habits for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_comitanigiacomo_kanso-sync-engine_internal_core_domain.Habit'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all habits
      tags:
      - Habits
    post:
      consumes:
      - application/json
      description: Create a habit with title, type, color, frequency, and tracking
        details
      parameters:
      - description: Habit Data
        in: body
        name: habit
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_handler_http.createHabitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_comitanigiacomo_kanso-sync-engine_internal_core_domain.Habit'
        "400":
          description: Validation Error (Title empty, Invalid Color)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new habit
      tags:
      - Habits
  /habits/{id}:
    delete:
      description: Mark a habit as deleted (archived)
      parameters:
      - description: Habit ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Habit Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Soft-delete a habit
      tags:
      - Habits
    put:
      consumes:
      - application/json
      description: Modify an existing habit. Requires 'version' for optimistic locking.
      parameters:
      - description: Habit ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Data
        in: body
        name: habit
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_handler_http.updateHabitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid Input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Habit Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Version Conflict (Data modified elsewhere)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a habit
      tags:
      - Habits
  /habits/sync:
    get:
      description: Get habits created, updated, or deleted since the provided timestamp
      parameters:
      - description: Timestamp (RFC3339 format)
        in: query
        name: last_sync
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'Returns {changes: delta, timestamp: now}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Timestamp Format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Sync habits (Offline-First)
      tags:
      - Habits
  /stats/weekly:
    get:
      description: 'Returns completion data for a specific date range (default: last
        7 days). Max range: 366 days.'
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns stats data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Date Format or Range > 1 Year
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get habit statistics
      tags:
      - Stats
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"